gmail-unsubscriber-backend/requirements-vercel.txt
flask==2.3.3
werkzeug==2.3.8
flask-cors==5.0.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.166.0
requests==2.31.0
beautifulsoup4==4.13.3
python-dotenv==1.1.0
PyJWT==2.10.1
anthropic==0.39.0
